+++ components/WeatherWindow/WeatherWindow.js
